Document Nomad setup and common tasks
This should be up to par with the same documentation we have on the old Aurora
cluster (i.e. setup of the cluster, local client setup, common tasks and how to
write job declarations).

The infrastructure-nomad page only documents the cluster itself; host
setup (i.e. analogous to the old infrastructure-mesos page) is documented on
infrastructure-nomad-host, to be submitted in a separate PR.